Convert to Decimal for voice speed

+++ src/main/java/org/ipg/namesvc/dto/EmployeeDTO.java
@@ -5,7 +5,7 @@ public record EmployeeDTO(
         String lastName,
         String preferredName,
         String preferredPreset,
-        Integer preferredSpeed,
+        Double preferredSpeed,
         boolean active) {
 
     public EmployeeDTO(String firstName, String lastName, String preferredName) {
